api/requirements.txt
argon2-cffi==23.1.0
cryptography==43.0.1
Flask==3.0.3
Flask-Cors==5.0.0
python-dotenv==1.0.1
requests==2.32.3
sentry-sdk==2.16.0
pyfcm==2.0.7
influxdb-client==1.48.0
